AER-4508 enhanced migration detection for scan
Because of AER-3639 being expanded to have migrations stats per
namespace we can now enhance the migration detection of a scan to only
disallow scans when the namespace being scanned is migrating. If other
namespaces are migrating but not the one being scanned then the scan may
proceed.

Also the new metrics are more robust than the old check which would
allow scans on a node that was still receiving migrations which means
data could be missing from the results.