AER-4656 Allow write proxy to proceed on reservation success
Regardless if the incoming proxies cluster key doesn't match the
local. This doesn't provide any real value and causes proxies to take
longer than necessary. Also eliminated the proxyee node from telling the
proxyer which node to go to next. The proxyer has had a half RTT longer
to learn who is the correct node and it may not be that node -- the old
system would have added an additional RTT delay.