AER-4877 Cluster fix for fast principal restart
On restart the sequence number used for paxos resets to one.

A restart of the principal node before other nodes in the cluster reform
a cluster has the principal stuck in a state where it sends sync
messages and other nodes reject the sync message because the pricipal
might have used a sequence number greater than one periously, but after
restart the sequence number used is one and sync fails The cluster
formation is then incomplete and stuck for ever until another principal is selected.

Fixed by

1. Adding a threshold for sync attempts after which auto-reset policy
comes into play instead of infinitely trying sync.
2. Using seconds from the wall clock for tne sequence number to prevent
sequence number going back to a previous value after a node
restart. With hbv3 this should be replaced with HLC clock.