Close out epoll, prevent access to closed socket
HB could close an fd while in the middle of handling a
batch of epoll events. Its possible that a subsequent epoll event in the
same epoll batch would be a read, this was ok since the older code dealt
with fds that altough invalid could still be referenced. The older code
would figure the fd is closed (since it was no longer in the channel
fd_to_channel hash and avoid a double close and any other access)

With the pointer change however the pointer itself will be freed and any
dereference on it (although not write would be incorrect).

Fixed this by performing the actual close out of the epoll loop to
prevent such races and avoid acessiong a freed up socket pointer.