AER-5309 Fix by relaxing paxos message rejection
In this particular case one node was rejecting legitimate paxos messages
from the new principal just because it was part of a cluster with the
older principal having greater nodeid.

This check makes sense only if the older principal is alive. Added this
additional is alive qualifier to relax the check and ensure cluster forms in such cases.